feat: Implement customer registration and login functionality with OAuth
- Updated OAuthController to handle customer registration and login.
- Created Customer model and controller for managing customer profiles.
- Added validation for customer data during registration.
- Implemented caching for customer and category data retrieval.
- Modified user role to 'customer' upon registration.
- Enhanced API routes for customer and admin functionalities.
- Added support for customer profile updates and retrieval.
- Implemented error handling for customer deletion with existing orders.
- Updated AuthServiceProvider to define access gates for admin and customer roles.
- Removed obsolete OAuth migrations and adjusted database structure for MongoDB.
- Updated composer dependencies for Laravel Passport.

+++ ecommerce-stock-management/config/oauth.php
@@ -31,7 +31,7 @@ return [
     |--------------------------------------------------------------------------
     */
     'tokens' => [
-        'access_token_lifetime' => env('OAUTH_ACCESS_TOKEN_LIFETIME', 3600), // 1 hour
-        'refresh_token_lifetime' => env('OAUTH_REFRESH_TOKEN_LIFETIME', 86400 * 30), // 30 days
+        'access_token_lifetime' => (int) env('OAUTH_ACCESS_TOKEN_LIFETIME', 3600), // Cast to int
+        'refresh_token_lifetime' => (int) env('OAUTH_REFRESH_TOKEN_LIFETIME', 86400 * 30), // Cast to int
     ],
 ];
